WiP - Improving the Domain Mapping Experience
We are adding a new card to the domain edit page to either provide
instructions on how to modify DNS records to map a domain to
WordPress.com or, in special cases, to automate the creation of the
appropriate DNS records at third party name server hosts.

This is still a work in progress and is not yet ready for any testing
outside of the team working on it.